requirements.txt
chromadb>=0.4.24
jq>=1.6.0
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1
langgraph>=0.0.40
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
sentence-transformers>=2.2.0  # For local embeddings
transformers>=4.21.0          # Required by sentence-transformers
torch>=1.12.0                 # Required by transformers (CPU version)

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
